Deployment/requirements.txt
tensorflow
numpy
pandas
scikit-learn
pillow
